+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Changelog
 
+- Added HTTP basic authentication.
 - Added the ability use local or remote state.
 
 ## v0.1.0 (2014-11-01)